+++ src/main/java/com/wanted/codebombalms/learning/application/port/LearningLectureProblemSetPort.java
@@ -3,6 +3,4 @@ package com.wanted.codebombalms.learning.application.port;
 public interface LearningLectureProblemSetPort {
 
     LearningLectureProblemSet findLectureProblemSet(Long lectureProblemSetId);
-
-    boolean existsProblemInSet(Long problemSetId, Long problemId);
 }
