+++ src/main/java/com/wanted/codebombalms/learning/application/port/LearningProblemPort.java
@@ -4,6 +4,8 @@ public interface LearningProblemPort {
 
     ProblemForLearning loadProblem(Long problemId);
 
+    boolean existsProblemInSet(Long problemSetId, Long problemId);
+
     record ProblemForLearning(
             Long problemId,
             String explanation,
